+++ MovieKing/src/main/java/com/spirng/movieKing/movie/service/MovieServiceImpl.java
@@ -1,10 +1,8 @@
-package com.spirng.movieList.controller;
+package com.spirng.movieKing.movie.service;
 
 import java.io.BufferedInputStream;
 import java.net.URL;
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 import org.json.simple.JSONArray;
@@ -12,27 +10,16 @@ import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.stereotype.Service;
 
-import com.spirng.movie.HomeController;
 
-@Controller
-@RequestMapping("/movies/*")
-public class MovieListController {
-	
-	private static final Logger logger = LoggerFactory.getLogger(MovieListController.class);
-	
-	@RequestMapping(value = "listJson", produces ="application/json;charset=UTF-8")
-	@ResponseBody
-	public JSONArray listJson() throws Exception {
+@Service
+public class MovieServiceImpl implements MovieService {
+	private static final Logger logger = LoggerFactory.getLogger(MovieServiceImpl.class);
+
+	@Override
+	public JSONArray getMovieList() throws Exception {
 		BufferedInputStream reader = null;
-		Map<String,Object> map=new HashMap<String, Object>();
 		try {
 			URL url = new URL("http://www.kobis.or.kr/kobisopenapi/webservice/rest/boxoffice/"
 					+ "searchDailyBoxOfficeList.json" + "?key=09ef5a1d8dc892084547547c61c96867" + "&targetDt=20181219"+"&itemPerPage=6");
@@ -57,16 +44,13 @@ public class MovieListController {
 				reader.close();
 		}
 	}
-	@RequestMapping(value="detailJson/{movieCd}",produces ="application/json;charset=UTF-8")
-	@ResponseBody
-	public Map<String,Object> getMovieJson(@PathVariable("movieCd")String movieCd) throws Exception{
+
+	@Override
+	public Map<String, Object> getMovie(String movieCd) throws Exception {
 		BufferedInputStream reader=null;
 		Map<String,Object> map=new HashMap<String, Object>();
-		List<String> directorsList=new ArrayList<String>();
-		List<String> list=new ArrayList<String>();
-	
 		
-		System.out.println(movieCd);
+	
 
 		try
 		
@@ -80,7 +64,6 @@ public class MovieListController {
 		while((i=reader.read(b))!=-1) {
 			buffer.append(new String(b,0,i));
 		}
-		System.out.println(buffer.toString());
 		JSONParser jsonparser = new JSONParser();
 		JSONObject jsonobject = (JSONObject) jsonparser.parse(buffer.toString());
 		JSONObject movieInfoResult = (JSONObject) jsonobject.get("movieInfoResult");
@@ -143,21 +126,6 @@ public class MovieListController {
 			if (reader != null)
 				reader.close();
 		}
-		
 	}
-	@RequestMapping(value = "/list")
-	public String list() {
-
-		return "movies/movie";
-	}
-	@RequestMapping(value="/detail")
-	public String detail(Model model,@RequestParam("movieCd")String movieCd,@RequestParam("im")String im,@RequestParam("audiCnt")String audiCnt) {
-		model.addAttribute("movieCd",movieCd);
-		model.addAttribute("im",im);
-		model.addAttribute("audiCnt",audiCnt);
-		return "movies/getMovie";
-	}
-	
-	
 
 }
