+++ MovieKing/src/main/resources/log4jdbc.log4j2.properties
@@ -1,0 +1,2 @@
+
+log4jdbc.spylogdelegator.name=net.sf.log4jdbc.log.slf4j.Slf4jSpyLogDelegator
