+++ MovieKing/src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
@@ -21,7 +21,7 @@
 		<beans:property name="suffix" value=".jsp" />
 	</beans:bean>
 	
-	<context:component-scan base-package="com.spirng.movie" />
-	<context:component-scan base-package="com.spirng.movieList" />
+	<context:component-scan base-package="com.spirng.movieKing" />
+
 
 </beans:beans>
