+++ MovieKing/src/main/webapp/WEB-INF/views/movies/nav.jsp
@@ -27,25 +27,26 @@
     <link href="/resources/css/creative.min.css" rel="stylesheet">
 	<!-- Google Map Key -->
  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js"></script>
-    
-    
+ <script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.18.0/moment.min.js"></script>
+ <script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.18.0/locale/ko.js"></script>
+  
   </head>
   <body id="page-top">
 
     <!-- Navigation -->
     <nav class="navbar navbar-expand-lg navbar-light fixed-top" id="mainNav" style="background-color:blue;">
       <div class="container">
-        <a class="navbar-brand js-scroll-trigger" href="#page-top">MovieKing</a>
+        <a class="navbar-brand js-scroll-trigger" href="/">MovieKing</a>
         <button class="navbar-toggler navbar-toggler-right" type="button" data-toggle="collapse" data-target="#navbarResponsive" aria-controls="navbarResponsive" aria-expanded="false" aria-label="Toggle navigation">
           <span class="navbar-toggler-icon"></span>
         </button>
         <div class="collapse navbar-collapse" id="navbarResponsive">
           <ul class="navbar-nav ml-auto">
             <li class="nav-item">
-              <a class="nav-link js-scroll-trigger" href="#about">영화</a>
+              <a class="nav-link js-scroll-trigger" href="/movies/list">영화</a>
             </li>
             <li class="nav-item">
-              <a class="nav-link js-scroll-trigger" href="#services">예매</a>
+              <a class="nav-link js-scroll-trigger" href="/service/">예매</a>
             </li>
             <li class="nav-item">
               <a class="nav-link js-scroll-trigger" href="#portfolio">로그인</a>
